+++ src/main/java/tahub/sdapitahub/controller/TaskCandidateController.java
@@ -11,6 +11,7 @@ import tahub.sdapitahub.entity.TaskCandidate;
 import tahub.sdapitahub.entity.TaskCandidateHistory;
 import tahub.sdapitahub.service.TaskCandidateHistoryService;
 import tahub.sdapitahub.service.TaskCandidateService;
+import tahub.sdapitahub.constants.TaskCandidateMsgs;
 
 import java.util.List;
 import java.util.Optional;
@@ -61,24 +62,45 @@ public class TaskCandidateController {
     }
 
     @PostMapping("/task-candidate")
-    public ResponseEntity<TaskCandidate> createTaskCandidate(@Valid @RequestBody TaskCandidateDTO taskCandidatePostDTO) {
+    public ResponseEntity<String> createTaskCandidate(@Valid @RequestBody TaskCandidateDTO taskCandidatePostDTO) {
         TaskCandidate createdTaskCandidate = taskCandidateService.createTaskCandidate(taskCandidatePostDTO);
-        return ResponseEntity.status(HttpStatus.CREATED).body(createdTaskCandidate);
-    }
-
-    @PutMapping("/task-candidate/{id}")
-    public ResponseEntity<TaskCandidate> updateTaskCandidate(@PathVariable Long id, @RequestBody TaskCandidateDTO taskCandidatePostDTO) {
-        TaskCandidate updatedTaskCandidate = taskCandidateService.updateTaskCandidate(id, taskCandidatePostDTO);
-        if (updatedTaskCandidate != null) {
-            return ResponseEntity.ok(updatedTaskCandidate);
+        if (createdTaskCandidate != null) {
+            return new ResponseEntity<String>(TaskCandidateMsgs.TASK_CANDIDATE_CREATED.getMessage(), HttpStatus.OK);
         } else {
-            return ResponseEntity.notFound().build();
+            return new ResponseEntity<String>(TaskCandidateMsgs.ERROR_TASK_CANDIDATE_CREATE.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
+    @PutMapping("/task-candidate/{id}")
+    public ResponseEntity<String> updateTaskCandidate(@PathVariable Long id, @RequestBody TaskCandidateDTO taskCandidatePostDTO) {
+        try {
+            TaskCandidate updatedTaskCandidate = taskCandidateService.updateTaskCandidate(id, taskCandidatePostDTO);
+            if (updatedTaskCandidate != null) {
+                return new ResponseEntity<String>(TaskCandidateMsgs.TASK_CANDIDATE_UPDATED.getMessage(), HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(TaskCandidateMsgs.TASK_CANDIDATE_NOT_FOUND.getMessage(), HttpStatus.NOT_FOUND);
+            }
+        }catch (Exception e){
+            return new ResponseEntity<String>(TaskCandidateMsgs.ERROR_TASK_CANDIDATE_UPDATE.getMessage(),HttpStatus.INTERNAL_SERVER_ERROR);
+
+        }
+
+    }
+
     @DeleteMapping("/task-candidate/{id}")
-    public ResponseEntity<Void> deleteTaskCandidate(@PathVariable Long id) {
-        taskCandidateService.deleteTaskCandidate(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<String> deleteTaskCandidate(@PathVariable Long id) {
+        try {
+            Optional<TaskCandidate> optionalTaskCandidate = taskCandidateService.getTaskCandidateById(id);
+            if (optionalTaskCandidate.isPresent()) {
+                taskCandidateService.deleteTaskCandidate(id);
+                return new ResponseEntity<String>(TaskCandidateMsgs.TASK_CANDIDATE_DELETED.getMessage(), HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity<>(TaskCandidateMsgs.TASK_CANDIDATE_NOT_FOUND.getMessage(), HttpStatus.NOT_FOUND);
+            }
+        }catch (Exception e){
+            return new ResponseEntity<String>(TaskCandidateMsgs.ERROR_TASK_CANDIDATE_DELETE.getMessage(),HttpStatus.INTERNAL_SERVER_ERROR);
+
+        }
     }
 }
